fix: update ventas

+++ src/main/java/com/gaspar/facturador/application/request/VentaSinFacturaRequest.java
@@ -27,5 +27,5 @@ public class VentaSinFacturaRequest {
     @NotNull(message = "El método de pago es obligatorio")
     private String metodoPago;
 
-    private Integer idfactura;
+    private Long idfactura;
 }
